item/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
bandit==1.7.4
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
pbr==5.9.0
Pillow==9.0.1
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
tzdata==2022.1
urllib3==1.26.8
